Move new utils/__init__.py funcs to utils.files.py
We're in the process of breaking up the utils/__init__.py file, so
let's steer clear of adding new funcs here.

This way we can avoid a deprecation process for these functions in
develop, since the funcs in this location were never released.

Refs #43056
